nubia/requirements/development.txt
-r production.txt

appnope==0.1.0
atomicwrites==1.3.0
attrs==19.3.0
backcall==0.1.0
decorator==4.4.1
importlib-metadata==0.23
ipdb==0.12.2
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
more-itertools==7.2.0
packaging==19.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.0
prompt-toolkit==2.0.10
ptyprocess==0.6.0
py==1.8.0
PyExecJS==1.5.1
Pygments==2.4.2
pyparsing==2.4.5
pytest==5.2.2
six==1.13.0
traitlets==4.3.3
wcwidth==0.1.7
zipp==0.6.0
